functions/instance-notification/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-asset
google-cloud-resource-manager
requests